+++ R/geom_edge_fan.R
@@ -77,12 +77,11 @@
 #'
 #' @examples
 #' require(tidygraph)
-#' gr <- as_tbl_graph(data.frame(
-#'   from = c(1, 1, 1, 1, 1, 2, 2, 2),
-#'   to = c(2, 2, 2, 2, 2, 1, 1, 1),
-#'   class = sample(letters[1:3], 8, TRUE)
-#' )) %>%
-#'   mutate(class = c('a', 'b'))
+#' gr <- create_notable('bull') %>%
+#'   convert(to_directed) %>%
+#'   bind_edges(data.frame(from = c(1, 2, 2, 3), to = c(2, 1, 3, 2))) %E>%
+#'   mutate(class = sample(letters[1:3], 9, TRUE)) %N>%
+#'   mutate(class = sample(c('x', 'y'), 5, TRUE))
 #'
 #' ggraph(gr, 'nicely') +
 #'   geom_edge_fan(aes(alpha = ..index..))