+++ NEWS.md
@@ -1,5 +1,6 @@
 # ggraph (development version)
 
+* Added `geom_edge_parallel()` for drawing multiedges as parallel lines (#191)
 * Added `geom_edge_bend()` for drawing soft elbows (#45)
 * Added `geom_node_voronoi()` for displaying nodes as voronoi tiles (#100)
 * All non-straight line-based edge geoms now has a `strength` parameter that